Versioning and calibration improvements
